move MDXComponents inside components folder
