+++ components/MDXComponents.js
@@ -3,12 +3,6 @@ import Image from "next/image";
 import { ExternalLinkIcon } from "@chakra-ui/icons";
 import { MdLabel } from "react-icons/md";
 import {
-  Accordion,
-  AccordionItem,
-  AccordionButton,
-  AccordionPanel,
-  AccordionIcon,
-  Box,
   Divider,
   Link,
   Heading,
@@ -25,14 +19,14 @@ import {
   Td,
 } from "@chakra-ui/react";
 
-import { Dropdown, DropdownItem } from "../Dropdown";
-import { BlogImage } from "./BlogImage";
+import { Dropdown, DropdownItem } from "./Dropdown";
+import { PostImage } from "./post/PostImage";
 import { YouTube } from "./YouTube";
 import { YouTubeBtn } from "./YouTubeBtn";
-import { TidycalEmbed } from "../TidycalEmbed";
-import { CompanyLogos } from "../icons/CompanyLogos";
-import { AboutCard } from "../cards/AboutCard";
-import { PageSummary } from "../page/PageSummary";
+import { TidycalEmbed } from "./TidycalEmbed";
+import { CompanyLogos } from "./icons/CompanyLogos";
+import { AboutCard } from "./cards/AboutCard";
+import { PageSummary } from "./page/PageSummary";
 
 const CustomLink = (props) => {
   const href = props.href;
@@ -59,14 +53,13 @@ const CustomLink = (props) => {
   );
 };
 
-const MDXComponents = {
+export const MDXComponents = {
   h1: (props) => <Heading as="h1" m={8} size="3xl" {...props} />,
   h2: (props) => <Heading as="h2" m={4} size="2xl" {...props} />,
   h3: (props) => <Heading as="h3" m={4} size="xl" {...props} />,
   h4: (props) => <Heading as="h4" m={2} size="lg" {...props} />,
   h5: (props) => <Heading as="h5" m={2} size="md" {...props} />,
   h6: (props) => <Heading as="h6" m={2} size="sm" {...props} />,
-  hr: Divider,
   ul: (props) => <UnorderedList styleType="none" {...props} />,
   ol: (props) => <OrderedList styleType="none" {...props} />,
   li: (props) => (
@@ -82,9 +75,10 @@ const MDXComponents = {
   th: (props) => <Th {...props} />,
   tr: (props) => <Tr {...props} />,
   td: (props) => <Td {...props} />,
-  Image,
-  img: (props) => <BlogImage {...props} />,
+  img: (props) => <PostImage {...props} />,
+  hr: Divider,
   a: CustomLink,
+  Image,
   PageSummary,
   Heading,
   YouTube,
@@ -95,5 +89,3 @@ const MDXComponents = {
   Dropdown,
   DropdownItem,
 };
-
-export default MDXComponents;
