+++ jsconfig.json
@@ -4,7 +4,6 @@
     "paths": {
       "@/layouts": ["layouts/"],
       "@/components": ["components/"],
-      "@/MDXComponents": ["components/MDX/"],
       "@/data/*": ["data/*"],
       "@/lib/*": ["lib/*"],
       "@/styles/*": ["styles/*"]
